refactor: update fee utility imports and add gwei definition
- Replaced local gwei imports with a centralized gwei definition in the new gwei.ts file for better code organization.
- Removed the KeysignSigningState component as it was no longer needed, streamlining the codebase.
- Updated related components to reflect the changes in imports and enhance maintainability.
