+++ clients/desktop/src/chain/tx/fee/utils/formatFee.ts
@@ -5,13 +5,13 @@ import { polkadotConfig } from '@core/chain/chains/polkadot/config'
 import { solanaConfig } from '@core/chain/chains/solana/solanaConfig'
 import { tonConfig } from '@core/chain/chains/ton/config'
 import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
+import { gwei } from '@core/chain/tx/fee/evm/gwei'
 import { rippleTxFee } from '@core/chain/tx/fee/ripple'
 import { KeysignChainSpecific } from '@core/mpc/keysign/chainSpecific/KeysignChainSpecific'
 import { isOneOf } from '@lib/utils/array/isOneOf'
 import { formatTokenAmount } from '@lib/utils/formatTokenAmount'
 import { matchDiscriminatedUnion } from '@lib/utils/matchDiscriminatedUnion'
 
-import { gwei } from './evm'
 import { getFeeUnit } from './feeUnit'
 
 type FormatFeeInput = {
