+++ clients/desktop/src/vault/keysign/shared/KeysignSigningStep.tsx
@@ -4,6 +4,7 @@ import { TxOverviewPanel } from '@core/ui/chain/tx/TxOverviewPanel'
 import { TxOverviewChainDataRow } from '@core/ui/chain/tx/TxOverviewRow'
 import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { KeysignCustomMessageInfo } from '@core/ui/mpc/keysign/custom/KeysignCustomMessageInfo'
+import { KeysignSigningState } from '@core/ui/mpc/keysign/flow/KeysignSigningState'
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { useVersion } from '@core/ui/product/state/version'
 import { Match } from '@lib/ui/base/Match'
@@ -21,7 +22,6 @@ import { extractErrorMsg } from '@lib/utils/error/extractErrorMsg'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { KeysignSigningState } from './KeysignSigningState'
 import { KeysignTxOverview } from './KeysignTxOverview'
 import { useKeysignMutation } from './mutations/useKeysignMutation'
 import { SwapKeysignTxOverview } from './SwapKeysignTxOverview'
